+++ js/main.js
@@ -1,14 +1,6 @@
-import './form-util.js';
-import './form-validation.js';
+import './form.js';
 import './slider.js';
 import './map.js';
-import { createDataErrorModal } from './form-modal.js';
-import { addMapFilterHandlers } from './map-filters.js';
-import { createPoints } from './map-util.js';
-import { getData } from './api.js';
-import './form-images.js';
+import { fillMap } from './data.js';
 
-getData((data) => {
-  createPoints(data);
-  addMapFilterHandlers(data);
-}, createDataErrorModal);
+fillMap();
